requirements.txt
Pillow~=9.1.1

argparse~=1.4.0